Split `admin` E2E group and merge it with `auditing`
