+++ e2e/test/scenarios/admin-2/auditing/approved-domains.cy.spec.js
