+++ e2e/test/scenarios/admin-2/auditing/auditing.cy.spec.js
