+++ e2e/test/scenarios/admin-2/auditing/subscriptions.cy.spec.js
