+++ e2e/test/scenarios/admin-2/people/reproductions/23689-sandboxed-group-manager.cy.spec.js
