+++ e2e/test/scenarios/admin-2/sso/jwt.cy.spec.js
@@ -10,7 +10,7 @@ import {
 import {
   crudGroupMappingsWidget,
   checkGroupConsistencyAfterDeletingMappings,
-} from "./group-mappings-widget";
+} from "./shared/group-mappings-widget";
 
 describeEE("scenarios > admin > settings > SSO > JWT", () => {
   beforeEach(() => {
