+++ e2e/test/scenarios/admin-2/sso/ldap.cy.spec.js
@@ -11,7 +11,7 @@ import {
 import {
   crudGroupMappingsWidget,
   checkGroupConsistencyAfterDeletingMappings,
-} from "./group-mappings-widget";
+} from "./shared/group-mappings-widget";
 
 describe(
   "scenarios > admin > settings > SSO > LDAP",
