+++ e2e/test/scenarios/admin-2/sso/shared/group-mappings-widget.js
