+++ e2e/test/scenarios/admin-2/whitelabel.cy.spec.js
